Working MVVLVA

+++ include/MoveGenerator.h
@@ -9,6 +9,6 @@
 
 extern std::vector<move_t> generateLegalMoves(Board& board);
 extern std::vector<move_t> generateMoves(Board& board);
-extern unsigned long long perft(Board& b, int depth, bool verbose = false);
+extern u64 perft(Board& b, int depth, bool verbose = false);
 
 #endif // MOVEGENERATOR_H_
